feat: 더미데이터에서 api 응답 데이터로 교체
