+++ src/components/mainPage/recommended/GroupProductCarousel.tsx
@@ -1,10 +1,18 @@
 import styled from 'styled-components';
 import { IoMdHeartEmpty } from 'react-icons/io';
-import { productData } from '../../../utils/productData';
 import { Link } from 'react-router-dom';
 import Slider from 'react-slick';
 import 'slick-carousel/slick/slick.css';
 import 'slick-carousel/slick/slick-theme.css';
+import axios from 'axios';
+import { API_URL } from '../../../constants/API_URL';
+import { useAppDispatch, useAppSelector } from '../../../app/hooks';
+import {
+  getGroupProductData,
+  groupData,
+} from '../../../features/main/groupSlice';
+import { useEffect } from 'react';
+import { ResponseType } from '../../../utils/ResponseType';
 
 const settings = {
   slidesToShow: 4,
@@ -13,20 +21,43 @@ const settings = {
 };
 
 export const GroupProductCarousel = () => {
+  const dispatch = useAppDispatch();
+  const products = useAppSelector(groupData);
+  const getGroupData: any = () => {
+    axios
+      .post<ResponseType>(API_URL.POST.MAIN, {
+        categoryList: ['그룹별상품'],
+      })
+      .then((res) => {
+        const mainData = res.data.data;
+        const groupData = mainData.mainCategoryPackageList.그룹별상품;
+        console.log('완료', groupData);
+        dispatch(getGroupProductData(groupData));
+        return groupData;
+      })
+      .catch(() => {
+        console.log('error');
+      });
+  };
+
+  useEffect(() => {
+    getGroupData();
+  }, []);
+
   return (
     <CarouselContainer {...settings}>
-      {productData.map((productData) => (
-        <Link to={productData.url} key={productData.id}>
+      {products.map((data: any) => (
+        <Link to={'offers/' + data.id} key={data.id}>
           <CardContainer>
-            <ProductWrap key={productData.id}>
-              <ProductImg src={productData.imageUrl} alt="product_image" />
+            <ProductWrap key={data.id}>
+              <ProductImg src={data.imageUrl} alt="product_image" />
               <ProductText>
-                <Title>{productData.title}</Title>
-                <Content>{productData.content}</Content>
-                <Price>{productData.price} 원</Price>
+                <Title>{data.title}</Title>
+                <Content>{data.summary}</Content>
+                <Price>{data.price} 원</Price>
               </ProductText>
               <ProductTag>
-                <TagTitle>{productData.tag}</TagTitle>
+                <TagTitle>{data.keyword}</TagTitle>
               </ProductTag>
               <IoMdHeartEmpty size="20px" className="wish-icon" />
             </ProductWrap>
@@ -70,7 +101,6 @@ const ProductWrap = styled.div`
   border-radius: 10px;
   border: 1px solid #ccc;
   position: relative;
-  height: 100%;
 
   &:hover {
     cursor: pointer;
@@ -88,8 +118,8 @@ const ProductWrap = styled.div`
 `;
 
 const ProductImg = styled.img`
-  width: 100%;
-  height: 56%;
+  width: 272px;
+  height: 180px;
 `;
 
 const ProductText = styled.div`
